+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 
 - It is now possible to delete Google Drive files from OSdatascanner. 
 
+- There's now an SBSYS-tab on Report. It's only visible for users with the right "sbsystab_access" setting/permission.
+
 ### General improvements
 
 - Adjusted responsive size of navigation buttons in the admin module.